added a modal form that displays score and takes in a name
Added the functionality for the modal in game_functions.js. Added styling to the modal and its elements in game_styles.css.  Made a slight change to how the maximum radius is calculated in welcome_function.js.  Added html code that will represent the modal and its contents in game.erb.