Minor changes and added skeleton for settings functionality
Added a get call option for '/settings' in driver.rb. Fixed a display bug for hiding the search name box when the user presses hide and added a function for loadSettings in welcome_functions.js. Fixed some inconsistent spacing in game.erb and welcome.erb. Added a settings.erb file.