Fixed a bug to prevent the user from avoiding the home page on first visit
Added a check in game_functions.js to see if the default color values were set, if not then the user would get redirected to the home page so they can be set. Added the same check to settings_functions.js. and made more comments. Made a few documentation additions to welcome_functions.js.